+++ modules/sagemaker/sagemaker-templates-service-catalog/templates/finetune_llm_evaluation/product_stack.py
@@ -31,6 +31,8 @@ class Product(servicecatalog.ProductStack):
         build_app_asset: s3_assets.Asset,
         pre_prod_account_id: str,
         prod_account_id: str,
+        sagemaker_domain_id: str,
+        sagemaker_domain_arn: str,
         **kwargs: Any,
     ) -> None:
         super().__init__(scope, id)
@@ -71,6 +73,10 @@ class Product(servicecatalog.ProductStack):
 
         Tags.of(self).add("sagemaker:project-id", sagemaker_project_id)
         Tags.of(self).add("sagemaker:project-name", sagemaker_project_name)
+        if sagemaker_domain_id:
+            Tags.of(self).add("sagemaker:domain-id", sagemaker_domain_id)
+        if sagemaker_domain_arn:
+            Tags.of(self).add("sagemaker:domain-arn", sagemaker_domain_arn)
 
         # create kms key to be used by the assets bucket
         kms_key = kms.Key(
@@ -235,6 +241,8 @@ class Product(servicecatalog.ProductStack):
             "build",
             project_name=sagemaker_project_name,
             project_id=sagemaker_project_id,
+            domain_id=sagemaker_domain_id,
+            domain_arn=sagemaker_domain_arn,
             model_package_group_name=model_package_group_name,
             model_bucket=model_bucket,
             pipeline_artifact_bucket=pipeline_artifact_bucket,