+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fix opensearch removal policy
 - update MWAA to 2.9.2
 - update mwaa constraints
+- limit length of id in model name to prevent model name becoming too long
 
 ## v1.2.0