+++ CHANGELOG.md
@@ -35,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - update mwaa constraints
 - limit length of id in model name to prevent model name becoming too long
 - add permission for get secret value in `hf_import_models` template
+- add manifests/tags parameters to one-click-template
 
 ## v1.2.0