+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### **Changed**
 
 - remove explicit module manifest account/region mappings from `fmops-qna-rag`
+- moved CI/CD infra to separate repository and added self mutation pipeline to provision infra for module `sagemaker-templates-service-catalog`
 - changed ECR encryption to KMS_MANAGED
 - changed encryption for each bucket to KMS_MANAGED
 - refactor `airflow-dags` module to use Pydantic
@@ -39,6 +40,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - update mwaa constraints
 - limit length of id in model name to prevent model name becoming too long
 - add permission for get secret value in `hf_import_models` template
+- add manifests/tags parameters to one-click-template
+- add integration tests for `mlflow-image`
 
 ## v1.2.0